requirements.txt
fastapi
uvicorn
aiohttp
asyncio
pyyaml
loguru
black
starlette
pymongo
aioschedule
requests